+++ package.json
@@ -8,7 +8,7 @@
     "start-server": "cd mhe_server && npm run start",
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
-    "test": "ng test --browsers=ChromeHeadless --watch=true --code-coverage",
+    "test": "ng test --code-coverage",
     "test:coverage": "ng test --browsers=ChromeHeadless --watch=false --code-coverage"
   },
   "private": true,
